Finalise features/channels imports/exports
