+++ features/channels/components/twitch-channel/index.ts
@@ -1,3 +1,2 @@
 export { TwitchChannelVideos } from "./TwitchChannelVideos";
-export type { VideoFilters } from "./TwitchChannelVideos";
 export { TwitchVideoListing } from "./TwitchVideoListing";
