+++ features/channels/index.ts
@@ -1,2 +1,3 @@
 export * from "./components/twitch-channel";
 export * from "./components/youtube-channel";
+export * from "./components/VideosFilterMenu";
