+++ features/channels/types/videoFilters.ts
@@ -1,0 +1,6 @@
+export interface VideoFilters {
+  dateFilter: Date;
+  minDurationFilter: number;
+  maxDurationFilter: number;
+  keywordFilter: string;
+}
